Bugfix in MySQL unit test

+++ tests/Storage/PDOMySQLStorageTest.php
@@ -14,12 +14,21 @@ final class PDOMySQLStorageTest extends PDOStorageTestCase {
 	const DB_USER = 'reef_test';
 	const DB_PASS = 'reef_test';
 	
+	private static $PDOException = null;
+	
 	public static function setUpBeforeClass() {
 		try {
 			static::$PDO = new \PDO("mysql:dbname=".static::DB_NAME.";host=".static::DB_HOST, static::DB_USER, static::DB_PASS);
 		} catch (PDOException $e) {
+			static::$PDO = null;
+			static::$PDOException = $e;
+		}
+	}
+	
+	public function setUp() {
+		if(empty(static::$PDO)) {
 			$this->markTestSkipped(
-				'Cannot connect to MySQL database ('.$e->getMessage().')'
+				'Could not connect to MySQL database ('.static::$PDOException->getMessage().')'
 			);
 		}
 	}
@@ -36,6 +45,10 @@ final class PDOMySQLStorageTest extends PDOStorageTestCase {
 	}
 	
 	public static function tearDownAfterClass() {
+		if(empty(static::$PDO)) {
+			return;
+		}
+		
 		// Make sure the test table is removed, even if some test has failed
 		$sth = static::$PDO->prepare("
 			DROP TABLE IF EXISTS test;
